+++ lib/src/components/components.dart
@@ -3,6 +3,7 @@ export 'badge/badge.dart';
 export 'button/button.dart';
 export 'card/card.dart';
 export 'chip/chip.dart';
+export 'overlay/overlay.dart';
 export 'skeleton/skeleton.dart';
 export 'state/state.dart';
 export 'textfield/textfield.dart';